+++ src/main/java/racingcar/Application.java
@@ -1,5 +1,9 @@
 package racingcar;
 
+import racingcar.game.Car;
+import racingcar.game.Game;
+import racingcar.tool.InputUtils;
+
 import java.util.List;
 
 public class Application {
